+++ src/containers/Home/Destinatari/Destinatari.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { Children, cloneElement } from 'react'
 import PropTypes from 'prop-types'
 import { Section, Row, Col } from 'react-materialize'
 import Container from '../../../components/Container/Container'
@@ -7,15 +7,15 @@ import Styles from './Destinatari.module.css'
 const Destinatari = ({ children, title, button }) => (
   <Section id="destinatari" className="grey lighten-4">
     <Container className="center ">
-      <h2 className="hide-on-large-only">{title}</h2>
+      {title ? <h2 className="hide-on-large-only">{title}</h2> : null}{' '}
       <Row className={Styles.Row}>
-        {React.Children.map(children, (child, index) => (
+        {Children.map(children, (child, index) => (
           <Col s={12} m={index === 0 ? 12 : 6} xl={4} className={Styles.Col}>
             {child}
           </Col>
         ))}
       </Row>
-      {React.cloneElement(button, { className: Styles.ContactBtn })}
+      {cloneElement(button, { className: Styles.ContactBtn })}
     </Container>
   </Section>
 )
