+++ src/containers/Home/Servizi/Servizi.js
@@ -1,18 +1,19 @@
-import React from 'react'
+import React, { cloneElement } from 'react'
 import PropTypes from 'prop-types'
 import { Section, Row } from 'react-materialize'
-import Container from '../../components/Container/Container'
+import Container from '../../../components/Container/Container'
+import Styles from './Servizi.module.css'
 
 const Servizi = ({ children, title, content, button }) => (
   <div id="aree-di-intervento">
     <Container className="center">
-      {title && <h2 className="hide-on-large-only blue-text text-darken-3">{title}</h2>}
-      {content && <p className="flow-text grey-text left-align">{content}</p>}
+      {title ? <h2 className="hide-on-large-only blue-text text-darken-3">{title}</h2> : null}{' '}
+      {content ? <p className="flow-text grey-text left-align">{content}</p> : null}{' '}
     </Container>
     <Section id="servizi" className="grey lighten-4">
       <Container className="center">
         <Row>{children}</Row>
-        {button}
+        {cloneElement(button, { className: Styles.Btn })}
       </Container>
     </Section>
   </div>
