+++ src/containers/Home/Servizi/Servizi.module.css
@@ -1,0 +1,26 @@
+.Btn {
+  & button {
+    align-items: center;
+    display: inline-flex;
+
+    &::before {
+      background: linear-gradient(to right, blue, ease-out, color(blue tint(15%)));
+      bottom: 0;
+      content: '';
+      left: 0;
+      position: absolute;
+      right: 0;
+      top: 0;
+      transform: scaleX(0);
+      transform-origin: 0 50%;
+      transition-duration: 0.5s;
+      transition-property: transform;
+      transition-timing-function: ease-out;
+      z-index: -1;
+    }
+
+    & svg {
+      transform: translateX(50%);
+    }
+  }
+}
