+++ src/libs/utils.js
@@ -17,9 +17,24 @@ const getImageSizesFromConfig = imageSizes => {
     .map(sizes => [parseInt(sizes[0]), parseInt(sizes[1])])
 }
 
+const scrollTo = (element = null, offset = 0, behavior = 'smooth') => {
+  if (element) {
+    window.scrollTo({
+      top: element.getBoundingClientRect().top - document.body.getBoundingClientRect().top - offset,
+      behavior: behavior,
+    })
+  } else {
+    window.scrollTo({
+      top: document.body.getBoundingClientRect().top - offset,
+      behavior: behavior,
+    })
+  }
+}
+
 module.exports = {
   getFilePath,
   getFileName,
   getFileExtension,
   getImageSizesFromConfig,
+  scrollTo,
 }
